Move existing tests for getRemoveProfileInformation
